Added font-awsome, and some layout changes

+++ app/controllers/categories_controller.rb
@@ -44,8 +44,10 @@ class CategoriesController < ApplicationController
   end
 
   def destroy
-    category.destroy
-    redirect_to categories_url, notice: 'Category was successfully destroyed.'
+    if current_user.admin?
+      category.destroy
+      redirect_to categories_url, notice: 'Category was successfully destroyed.'
+    end
   end
 
   private
